requirements.txt
fastapi~=0.88.0
uvicorn~=0.20.0
pydantic~=1.10.4
SQLAlchemy~=1.4.45
passlib[bcrypt]~=1.7.4
python-jose[cryptography]~=3.3.0
psycopg2-binary~=2.9.5
python-dotenv~=0.21.0